style: rename files and functions
