+++ src/app/api/_helpers/createLog.js
@@ -1,5 +1,5 @@
 import { auth } from "@/src/auth";
-import { currentTime } from "@/src/services/moment";
+import { currentTime } from "@/src/services/timeUtils";
 
 export async function createLog({
   db,
