+++ src/services/api/parameter.js
@@ -1,8 +1,8 @@
-import { handleRequest } from "../apiRequest";
+import { handleRequest } from "../apiUtils";
 
 const apiRequest = handleRequest("Failed to fetch parameter");
 
-export async function getModuleOptions() {
-  const url = "/api/param/module/option";
+export async function getSystemModuleOptions() {
+  const url = "/api/param/system_modules/option";
   return apiRequest(url, { cache: "no-store" });
 }
